LayoutTests/js/typedarray-fill-expected.txt

This test checks the behavior of the TypedArray.prototype.fill function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.fill.length is 1
PASS Int32Array.prototype.fill.name is 'fill'
PASS isSameFunctionForEachTypedArrayPrototype('fill') is true
PASS testPrototypeReceivesArray('fill', [undefined, this, { }, [ ], true, '']) is true

testPrototypeFunction has the following arg list (name, args, init, result [ , expectedArray ])

1.0 Single Argument Testing
PASS testPrototypeFunction('fill', '(12)', [15, 5, 8, 13, 44], [12,12,12,12,12]) is true
PASS testPrototypeFunction('fill', '(true)', [12, 54, 18, 13, 44], [1,1,1,1,1]) is true

2.0 Two Argument Testing
PASS testPrototypeFunction('fill', '(12, 2)', [14, 15, 10, 13, 44], [14, 15, 12, 12, 12]) is true
PASS testPrototypeFunction('fill', '(4, NaN)', [14, 15, 10, 13, 44], [4, 4, 4, 4, 4]) is true
PASS testPrototypeFunction('fill', '(4, -5)', [14, 15, 10, 13, 44], [4, 4, 4, 4, 4]) is true
PASS testPrototypeFunction('fill', '(4, -1)', [14, 15, 10, 13, 44], [14, 15, 10, 13, 4]) is true

3.0 Three Argument Testing
PASS testPrototypeFunction('fill', '(4, -1, 0)', [14, 15, 10, 13, 44], [14, 15, 10, 13, 44]) is true
PASS testPrototypeFunction('fill', '(4, 1, 1)', [14, 15, 10, 13, 44], [14, 15, 10, 13, 44]) is true
PASS testPrototypeFunction('fill', '(4, 1, NaN)', [14, 15, 10, 13, 44], [14, 15, 10, 13, 44]) is true
PASS testPrototypeFunction('fill', '(4, NaN, NaN)', [14, 15, 10, 13, 44], [14, 15, 10, 13, 44]) is true
PASS testPrototypeFunction('fill', '(4, NaN, 5)', [14, 15, 10, 13, 44], [4, 4, 4, 4, 4]) is true
PASS testPrototypeFunction('fill', '(4, -3, -2)', [14, 15, 10, 13, 44], [14, 15, 4, 13, 44]) is true
PASS testPrototypeFunction('fill', '(4, 5, 5)', [14, 15, 10, 13, 44], [14, 15, 10, 13, 44]) is true
PASS successfullyParsed is true

TEST COMPLETE
